% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reliability.R
\name{coef_alpha}
\alias{coef_alpha}
\title{Coefficient Alpha}
\usage{
coef_alpha(data)
}
\arguments{
\item{data}{A matrix or data.frame containing the item level responses}
}
\description{
This calculates coefficient alpha for a given matrix or data.frame. Note that the matrix or data.frame should be at the item level and that it needs to be numeric not characters or factors.
}

