% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reliability.R
\name{kr20}
\alias{kr20}
\title{Kuder-Richardson 20}
\usage{
kr20(data)
}
\arguments{
\item{data}{A matrix or data.frame containing the item level responses}
}
\description{
This calculates Kuder-Richarson 20 for a given matrix or data.frame. Note that the matrix or data.frame should be at the item level and that it needs to be numeric not characters or factors. KR-20 is appropriate for dichotomously scored items, \code{coef_alpha} is appropriate for non-dichotomous items.
}

