% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reliability.R
\name{new_length}
\alias{new_length}
\title{New test length given a desired reliability}
\usage{
new_length(desired, present, test_length)
}
\arguments{
\item{desired}{Desired reliability}

\item{present}{Current reliability}

\item{test_length}{Current test length}
}
\description{
The calculates what the length of a test should be given your desired reliability, your present reliability, and the current length of the test.
}

