% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reliability.R
\name{sem}
\alias{sem}
\title{Standard error of measurement}
\usage{
sem(sigma, reliability)
}
\arguments{
\item{sigma}{Standard deviation of total test scores}

\item{reliablity}{Current reliability}
}
\description{
This calculates the standard error of measurement given the standard deviation of your total scores and your current reliabilty
}

