% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reliability.R
\name{spear_half}
\alias{spear_half}
\title{Spearman Brown Correction: Split - Half}
\usage{
spear_half(x)
}
\arguments{
\item{x}{The correlation of the two split halves}
}
\description{
This calculates the Spearman Brown correction for a split half only.
}

